spline #684 Call Foxx service
