+++ persistence/pom.xml
@@ -65,7 +65,7 @@
         <dependency>
             <groupId>com.arangodb</groupId>
             <artifactId>arangodb-java-driver</artifactId>
-            <version>6.6.3</version>
+            <version>6.7.3</version>
         </dependency>
         <dependency>
             <groupId>com.arangodb</groupId>
